Refactor ETL files to break them down into smaller folders and files
